Add test for map
