+++ algorithms/map/__init__.py
@@ -1,0 +1,2 @@
+from .hashtable import *
+from .separate_chaining_hashtable import *
